parquet: updated parquet-go dependency to v1.6.2 for not needing CGO_ENABLED=1 anymore;

+++ pkg/parquet/reader.go
@@ -274,7 +274,7 @@ func (r *s3Reader) ReadFileColumns(ctx context.Context, columnNames []string, fi
 			continue
 		}
 
-		values, _, _, err := pr.ReadColumnByPath(schema, size)
+		values, _, _, err := pr.ReadColumnByPath(schema, int64(size))
 		if err != nil {
 			return nil, err
 		}
